changes after comments
