+++ tests/cypress/e2e/menuActions/export.cy.ts
@@ -1,9 +1,40 @@
 import {JContent} from '../../page-object';
-import {createSite, deleteSite} from '@jahia/cypress';
+import {createSite, deleteSite, getComponent} from '@jahia/cypress';
+import {ExportDialog} from '../../page-object/exportDialog';
 
 describe('Export tests', () => {
     const siteKey = 'jContentSite-export';
     const downloadsFolder = Cypress.config('downloadsFolder');
+
+    function exportAndReimportAsZip(character: string) {
+        cy.login();
+        const labelWithSpecialCharacters = 'Special' + character + 'characters';
+
+        cy.apollo({
+            mutationFile: 'jcontent/createTextContentUnderPath.graphql',
+            variables: {contentName: labelWithSpecialCharacters, path: `/sites/${siteKey}/contents`}
+        });
+
+        const jcontent = JContent.visit(siteKey, 'en', 'content-folders/content');
+
+        jcontent.getTable().getRowByLabel(labelWithSpecialCharacters).contextMenu().select('Export');
+        const exportDialog = getComponent(ExportDialog);
+        exportDialog.export(downloadsFolder);
+        cy.waitUntil(() => cy.exec(`ls ${downloadsFolder}`).then(result => {
+            console.log(result.stdout);
+            return result.stdout.includes(`${labelWithSpecialCharacters}.zip`);
+        }), {timeout: 30000, interval: 1000, errorMsg: 'Unable to download content as zip'});
+
+        // Reimport exported content
+        jcontent.getTable().getRowByLabel('importFolder').get().dblclick();
+        jcontent.startImport();
+
+        cy.get('#file-upload-input').selectFile(`${downloadsFolder}/${labelWithSpecialCharacters}.zip`, {force: true});
+        jcontent.getTable().getRowByLabel(labelWithSpecialCharacters).get().dblclick();
+        jcontent.getTable().getRowByLabel('test path 1').should('exist');
+        jcontent.getTable().getRowByLabel('test path 2').should('exist');
+        jcontent.getTable().getRowByLabel('test path 3').should('exist');
+    }
 
     before(() => {
         cy.login();
@@ -46,27 +77,4 @@ describe('Export tests', () => {
     it('should export a content with characters space as zip', () => {
         exportAndReimportAsZip(' ');
     });
-
-    function exportAndReimportAsZip(character: string) {
-        cy.login();
-        const labelWithSpecialCharacters = 'Special' + character + 'characters';
-
-        cy.apollo({
-            mutationFile: 'jcontent/createTextContentUnderPath.graphql',
-            variables: {contentName: labelWithSpecialCharacters, path: `/sites/${siteKey}/contents`}
-        });
-
-        const jcontent = JContent.visit(siteKey, 'en', 'content-folders/content');
-
-        jcontent.exportComponentByText(labelWithSpecialCharacters, downloadsFolder, 'Staging content only', 'zip');
-
-        // Reimport exported content
-        jcontent.getTable().getRowByLabel('importFolder').get().dblclick();
-        jcontent.getBrowseControlMenu().selectByRole('import');
-        cy.get('#file-upload-input').selectFile(`${downloadsFolder}/${labelWithSpecialCharacters}.zip`, {force: true});
-        jcontent.getTable().getRowByLabel(labelWithSpecialCharacters).get().dblclick();
-        jcontent.getTable().getRowByLabel('test path 1').should('exist');
-        jcontent.getTable().getRowByLabel('test path 2').should('exist');
-        jcontent.getTable().getRowByLabel('test path 3').should('exist');
-    }
 });
